extracted Vuex fro reactions to separete file
