+++ resources/assets/js/components/qna/Vote.vue
@@ -74,5 +74,4 @@
 			},
 		},
 	}
-
 </script>
